extract BNF attribute API from generator to BnfAttributes
ParserGeneratorUtil.{getAttribute, getRootAttribute, findAttribute,
getAttributeValue, getLiteralValue, getTokenValue} are thin wrappers around
BnfFile.findAttribute/findAttributeValue plus literal-extraction helpers.
None of them are code generation — they're how anything reads BNF
attributes. Move to a new top-level org.intellij.grammar.psi.BnfAttributes.

Continues the prep for extracting :generator: PSI/analysis classes
(BnfReferenceImpl, BnfFileImpl, JavaHelper, GrammarUtil, GrammarPsiImplUtil,
BnfFirstNextAnalyzer) no longer need ParserGeneratorUtil for attribute reads.
