+++ src/org/intellij/grammar/livePreview/LivePreviewLexer.java
@@ -22,7 +22,7 @@ import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static org.intellij.grammar.generator.ParserGeneratorUtil.getRootAttribute;
+import static org.intellij.grammar.psi.BnfAttributes.getRootAttribute;
 import static org.intellij.grammar.livePreview.LivePreviewParserDefinition.*;
 
 /**
